+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - XXXX-XX-XX
+
+### Added
+- `FilePass`, `TypeDeclPass`, `PackagePass`
+
+### Changes
+- Graph builders are now known as "passes" to conform to how SCPG builds graphs
+
 ## [0.2.8] - 2021-03-05
 
 ### Added